+++ loraham_kiss_tnc/build.sh
@@ -59,6 +59,7 @@ sources=(
   bridge_conf.c
   bridge_kiss.c
   bridge_rx.c
+  bridge_loraham.c
   bridge_runtime.c
   bridge_tx_queue.c
   cli.c
